requirements.txt
django==3.2
djangorestframework==3.12
django-filter==2.4
gunicorn==20.0.4
langchain
langchain[all]
langchain-community
langsmith
beautifulsoup4
requests
celery==5.2.0
django-celery-beat==2.2.1
django-environ
psycopg2-binary